Fixing the simulator
